third_party/WebKit/LayoutTests/svg/parser/whitespace-angle-2.html

<!doctype html> 
<title>Whitespace in <angle> attribute values</title>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script src=resources/whitespace-helper.js></script>
<svg id="testcontainer">
	<defs>
		<marker/>
	</defs>
</svg>
<div id=log></div>
<script>

// test length values
var EPSILON = Math.pow(2, -24); // float epsilon
var whitespace = [ "", " ", "   ", "\r\n\t ", "\f" ];
var garbage = [ "a", "e", "foo", ")90" ];

// Identical to whitespace-angle-1 apart from valid entries. Split to
// reduce test timeouts on Linux Debug.
testType("<angle>",
		 document.querySelector("marker"),
		 "orient",
		 0, // expected default value
		 whitespace,
		 [ "+32", "+17E-1", "17e+2" ], // valid
		 [ "", "deg", "rad", "grad" ], // valid units
		 garbage,
		 function(elm, value) { assert_approx_equals(elm.orientAngle.baseVal.valueInSpecifiedUnits, parseFloat(value), EPSILON); },
		 function(elm, expected) { assert_approx_equals(elm.orientAngle.baseVal.value, expected, EPSILON); } );

</script>
