third_party/blink/web_tests/svg/parser/whitespace-angle-2-expected.txt

CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32 deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1 deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2 deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32 rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1 rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2 rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32 grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1 grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2 grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +32a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +17E-1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 17e+2a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +32e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +17E-1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 17e+2e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +32foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +17E-1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 17e+2foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +32)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +17E-1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " 17e+2)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32   deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1   deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2   deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32   rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1   rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2   rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32   grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1   grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2   grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +32a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +17E-1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   17e+2a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +32e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +17E-1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   17e+2e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +32foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +17E-1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   17e+2foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +32)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +17E-1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   17e+2)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32\r\n\t deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1\r\n\t deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2\r\n\t deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32\r\n\t rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1\r\n\t rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2\r\n\t rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32\r\n\t grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1\r\n\t grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2\r\n\t grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +32a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +17E-1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 17e+2a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +32e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +17E-1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 17e+2e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +32foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +17E-1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 17e+2foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +32)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +17E-1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t 17e+2)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32\fdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1\fdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2\fdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32\frad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1\frad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2\frad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+32\fgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+17E-1\fgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "17e+2\fgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+32a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+17E-1a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f17e+2a".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+32e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+17E-1e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f17e+2e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+32foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+17E-1foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f17e+2foo".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+32)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+17E-1)90".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f17e+2)90".
This is a testharness.js-based test.
Harness: the test ran to completion.
